Initial working version as a cli tool.
