+++ glide.yaml
@@ -1,3 +1,4 @@
 package: j2xrp
 import:
 - package: github.com/clbanning/mxj
+- package: github.com/sirupsen/logrus
